Simple Diffs without cancellative properties
We might be able to use Diffs without cancellative properties in some
parts of post-UTXO-HD `ouroboros-consensus`, opening up an opportunity
for optimising cases where we're not subtracting diffs from one another.
We can make the conversions between the anti-Diffs and the simple diffs,
but information might be lost in the process.